refactoring code validation params middleware
