+++ .gitignore
@@ -6,4 +6,3 @@ forgot_password.html
 reset_password.html
 login.html
 registration.html
-validate.params.ts
